Refactor MCP tool schemas
