+++ crates/rulemorph_mcp/src/main.rs
@@ -18,10 +18,16 @@ use serde_yaml::{Mapping as YamlMapping, Value as YamlValue};
 mod prompts;
 mod protocol;
 mod resources;
+mod schemas;
 
 use self::prompts::{prompts_get_result, prompts_list_result};
 use self::protocol::{OutputMode, read_message, write_message};
 use self::resources::{resources_list_result, resources_read_result};
+use self::schemas::{
+    analyze_input_input_schema, generate_dto_input_schema, generate_rules_from_base_input_schema,
+    generate_rules_from_dto_input_schema, list_ops_input_schema, transform_input_schema,
+    validate_rules_input_schema,
+};
 
 const PROTOCOL_VERSION: &str = "2024-11-05";
 const MCP_MAX_FILE_BYTES: usize = 64 * 1024 * 1024;
@@ -192,297 +198,6 @@ fn tools_list_result() -> Value {
                 "inputSchema": generate_rules_from_dto_input_schema()
             }
         ]
-    })
-}
-
-fn transform_input_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "rules_path": {
-                "type": "string",
-                "description": "Path to the YAML or JSON rules file. Mutually exclusive with rules_text.",
-                "examples": ["rules.yaml"]
-            },
-            "rules_text": {
-                "type": "string",
-                "description": "Inline YAML or JSON rules content. Mutually exclusive with rules_path.",
-                "examples": ["version: 1\ninput:\n  format: json\n  json: {}\nmappings:\n  - target: \"id\"\n    source: \"id\""]
-            },
-            "rules_format": rules_format_schema(),
-            "input_path": {
-                "type": "string",
-                "description": "Path to the input file. Mutually exclusive with input_text and input_json.",
-                "examples": ["input.json"]
-            },
-            "input_text": {
-                "type": "string",
-                "description": "Inline text input. Mutually exclusive with input_path and input_json.",
-                "examples": ["{\"items\":[{\"id\":1}]}"]
-            },
-            "input_json": {
-                "type": ["object", "array"],
-                "description": "Inline typed JSON value. Mutually exclusive with input_path and input_text. Duplicate-key validation applies to raw JSON input_text/input_path, not to input_json after JSON-RPC decoding.",
-                "examples": [[{"id": 1}]]
-            },
-            "context_path": {
-                "type": "string",
-                "description": "Optional path to a JSON context file. Mutually exclusive with context_json.",
-                "examples": ["context.json"]
-            },
-            "context_json": {
-                "type": "object",
-                "description": "Optional inline JSON context value. Mutually exclusive with context_path.",
-                "examples": [{"tenant_id": "t-001"}]
-            },
-            "format": {
-                "type": "string",
-                "enum": ["csv", "json", "yaml", "toml", "xml", "html", "excel"],
-                "description": "Override input format from the rule file.",
-                "examples": ["json"]
-            },
-            "ndjson": {
-                "type": "boolean",
-                "description": "Emit NDJSON output (one JSON object per line).",
-                "examples": [false]
-            },
-            "validate": {
-                "type": "boolean",
-                "description": "Validate the rule file before transforming.",
-                "examples": [true]
-            },
-            "output_path": {
-                "type": "string",
-                "description": "Optional path to write the output.",
-                "examples": ["out.json"]
-            },
-            "max_output_bytes": {
-                "type": "integer",
-                "minimum": 1,
-                "description": "Maximum output size in bytes before truncation.",
-                "examples": [1000000]
-            },
-            "preview_rows": {
-                "type": "integer",
-                "minimum": 1,
-                "description": "Maximum rows to return when ndjson=true.",
-                "examples": [100]
-            },
-            "return_output_json": {
-                "type": "boolean",
-                "description": "Include parsed output JSON in meta.output when ndjson=false and within size limits.",
-                "examples": [false]
-            }
-        }
-    })
-}
-
-fn rules_format_schema() -> Value {
-    json!({
-        "type": "string",
-        "enum": ["yaml", "json"],
-        "description": "Rule parser format. Defaults to file extension for rules_path and yaml for rules_text.",
-        "examples": ["json"]
-    })
-}
-
-fn validate_rules_input_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "rules_path": {
-                "type": "string",
-                "description": "Path to the YAML or JSON rules file. Mutually exclusive with rules_text.",
-                "examples": ["rules.yaml"]
-            },
-            "rules_text": {
-                "type": "string",
-                "description": "Inline YAML or JSON rules content. Mutually exclusive with rules_path.",
-                "examples": ["version: 1\ninput:\n  format: json\n  json: {}\nmappings:\n  - target: \"id\"\n    source: \"id\""]
-            },
-            "rules_format": rules_format_schema()
-        }
-    })
-}
-
-fn generate_dto_input_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "rules_path": {
-                "type": "string",
-                "description": "Path to the YAML or JSON rules file. Mutually exclusive with rules_text.",
-                "examples": ["rules.yaml"]
-            },
-            "rules_text": {
-                "type": "string",
-                "description": "Inline YAML or JSON rules content. Mutually exclusive with rules_path.",
-                "examples": ["version: 1\ninput:\n  format: json\n  json: {}\nmappings:\n  - target: \"id\"\n    source: \"id\""]
-            },
-            "rules_format": rules_format_schema(),
-            "language": {
-                "type": "string",
-                "enum": ["rust", "typescript", "python", "go", "java", "kotlin", "swift"],
-                "description": "DTO output language.",
-                "examples": ["typescript"]
-            },
-            "name": {
-                "type": "string",
-                "description": "Optional DTO root type name.",
-                "examples": ["Record"]
-            }
-        },
-        "required": ["language"]
-    })
-}
-
-fn list_ops_input_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {}
-    })
-}
-
-fn analyze_input_input_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "input_path": {
-                "type": "string",
-                "description": "Path to the input CSV/JSON file. Mutually exclusive with input_text and input_json.",
-                "examples": ["input.json"]
-            },
-            "input_text": {
-                "type": "string",
-                "description": "Inline input text (CSV or JSON). Mutually exclusive with input_path and input_json.",
-                "examples": ["{\"items\":[{\"id\":1}]}"]
-            },
-            "input_json": {
-                "type": ["object", "array"],
-                "description": "Inline typed JSON value. Mutually exclusive with input_path and input_text. Duplicate-key validation applies to raw JSON input_text/input_path, not to input_json after JSON-RPC decoding.",
-                "examples": [[{"id": 1}]]
-            },
-            "format": {
-                "type": "string",
-                "enum": ["csv", "json"],
-                "description": "Input format when input_text/input_path is used.",
-                "examples": ["json"]
-            },
-            "records_path": {
-                "type": "string",
-                "description": "Optional records path for JSON inputs.",
-                "examples": ["items"]
-            },
-            "max_paths": {
-                "type": "integer",
-                "minimum": 1,
-                "description": "Maximum number of unique paths to include in the response.",
-                "examples": [200]
-            }
-        }
-    })
-}
-
-fn generate_rules_from_base_input_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "rules_path": {
-                "type": "string",
-                "description": "Path to the YAML or JSON rules file. Mutually exclusive with rules_text.",
-                "examples": ["rules.yaml"]
-            },
-            "rules_text": {
-                "type": "string",
-                "description": "Inline YAML or JSON rules content. Mutually exclusive with rules_path.",
-                "examples": ["version: 1\ninput:\n  format: json\n  json: {}\nmappings:\n  - target: \"id\"\n    source: \"id\""]
-            },
-            "rules_format": rules_format_schema(),
-            "input_path": {
-                "type": "string",
-                "description": "Path to the input CSV/JSON file. Mutually exclusive with input_text and input_json.",
-                "examples": ["input.json"]
-            },
-            "input_text": {
-                "type": "string",
-                "description": "Inline input text (CSV or JSON). Mutually exclusive with input_path and input_json.",
-                "examples": ["{\"items\":[{\"id\":1}]}"]
-            },
-            "input_json": {
-                "type": ["object", "array"],
-                "description": "Inline typed JSON value. Mutually exclusive with input_path and input_text. Duplicate-key validation applies to raw JSON input_text/input_path, not to input_json after JSON-RPC decoding.",
-                "examples": [[{"id": 1}]]
-            },
-            "format": {
-                "type": "string",
-                "enum": ["csv", "json"],
-                "description": "Override input format.",
-                "examples": ["json"]
-            },
-            "records_path": {
-                "type": "string",
-                "description": "Optional records path for JSON inputs.",
-                "examples": ["items"]
-            },
-            "max_candidates": {
-                "type": "integer",
-                "minimum": 1,
-                "description": "Maximum number of candidates to return per target.",
-                "examples": [3]
-            }
-        }
-    })
-}
-
-fn generate_rules_from_dto_input_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "dto_text": {
-                "type": "string",
-                "description": "DTO source text.",
-                "examples": ["export interface Record { id: string; }"]
-            },
-            "dto_language": {
-                "type": "string",
-                "enum": ["rust", "typescript", "python", "go", "java", "kotlin", "swift"],
-                "description": "DTO language.",
-                "examples": ["typescript"]
-            },
-            "input_path": {
-                "type": "string",
-                "description": "Path to the input CSV/JSON file. Mutually exclusive with input_text and input_json.",
-                "examples": ["input.json"]
-            },
-            "input_text": {
-                "type": "string",
-                "description": "Inline input text (CSV or JSON). Mutually exclusive with input_path and input_json.",
-                "examples": ["{\"items\":[{\"id\":1}]}"]
-            },
-            "input_json": {
-                "type": ["object", "array"],
-                "description": "Inline typed JSON value. Mutually exclusive with input_path and input_text. Duplicate-key validation applies to raw JSON input_text/input_path, not to input_json after JSON-RPC decoding.",
-                "examples": [[{"id": 1}]]
-            },
-            "format": {
-                "type": "string",
-                "enum": ["csv", "json"],
-                "description": "Override input format.",
-                "examples": ["json"]
-            },
-            "records_path": {
-                "type": "string",
-                "description": "Optional records path for JSON inputs.",
-                "examples": ["items"]
-            },
-            "max_candidates": {
-                "type": "integer",
-                "minimum": 1,
-                "description": "Maximum number of candidates to return per target.",
-                "examples": [3]
-            }
-        },
-        "required": ["dto_text", "dto_language"]
     })
 }
